+++ backend/app-orchestrator-service/api/src/main/kotlin/NetworkIPs.kt
@@ -194,6 +194,7 @@ data class NetworkIPFlags(
 data class NetworkIPSupport(
     override val product: ProductReference,
     val firewall: Firewall = Firewall(),
+    override var maintenance: Maintenance? = null,
 ) : ProductSupport {
     @Serializable
     @UCloudApiStable