+++ backend/file-ucloud-service/src/main/resources/db/migration/V14__QuotaLockCollection.sql
@@ -1,5 +1,0 @@
-delete from file_ucloud.quota_locked where true;
-alter table file_ucloud.quota_locked drop column collection;
-alter table file_ucloud.quota_locked add column category text not null default '';
-alter table file_ucloud.quota_locked add column username text default null;
-alter table file_ucloud.quota_locked add column project_id text default null;